+++ src/main/kotlin/kr/craft/javaboilercraft/completion/impl/restdoc/core/HttpMethodMapping.kt
@@ -1,4 +1,4 @@
-package kr.craft.javaboilercraft.processor.impl.restdoc.core
+package kr.craft.javaboilercraft.completion.impl.restdoc.core
 
 /**
  * HttpMethodMapping
